+++ packages/manager/modules/billing/src/billing.routing.js
@@ -1,5 +1,9 @@
 import controller from './billing.controller';
 import template from './billing.html';
+import {
+  GUIDES_LIST,
+  GUIDE_TRACKING_TAG,
+} from './constants/guides-header.constants';
 
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('app.account.billing', {
@@ -25,6 +29,53 @@ export default /* @ngInject */ ($stateProvider) => {
       },
       goToOrders: /* @ngInject */ ($state) => () =>
         $state.go('app.account.billing.orders'),
+
+      guides: /* @ngInject */ (guidesObjectLevelOne) => {
+        return {
+          url: guidesObjectLevelOne(GUIDES_LIST),
+          tracking: GUIDE_TRACKING_TAG,
+        };
+      },
+
+      guidesObjectLevelOne: /* @ngInject */ (guidesObjectLevelTwo) => (
+        object,
+      ) => {
+        return Object.fromEntries(
+          Object.entries(object).map(([key, value]) => {
+            return [key, guidesObjectLevelTwo(value)];
+          }),
+        );
+      },
+
+      guidesObjectLevelTwo: /* @ngInject */ (guidesObjectLevelThree) => (
+        object,
+      ) => {
+        return Object.fromEntries(
+          Object.entries(object).map(([key, value]) => {
+            return [key, guidesObjectLevelThree(value)];
+          }),
+        );
+      },
+
+      guidesObjectLevelThree: /* @ngInject */ (currentUser) => (object) => {
+        return Object.fromEntries(
+          Object.entries(object).map(([key, value]) => {
+            return [
+              key,
+              key === 'url'
+                ? value[currentUser.ovhSubsidiary] || value.DEFAULT
+                : value,
+            ];
+          }),
+        );
+      },
+
+      trackClick: /* @ngInject */ (atInternet) => (hit) => {
+        return atInternet.trackClick({
+          name: hit,
+          type: 'action',
+        });
+      },
     },
   });
 };